+++ Totem_2022/Totem_2022.h
@@ -1,9 +1,9 @@
 #ifndef MAIN_H
 #define MAIN_H
 
-extern uint16_t effectMode;   // Chase, Fade
-extern uint16_t colorScheme;  // Rainbow, Pairs
-extern uint16_t brightness;   // 0 to 255
+extern uint8_t effectMode;    // Chase, Fade
+extern uint8_t colorScheme;   // Rainbow, Pairs
+extern uint8_t brightness;    // 0 to 255
 extern uint16_t effectSpeed;  // 0 to 1024
 
 extern bool tapEnabled;
