+++ Totem_2022/Totem_2022.ino
@@ -1,12 +1,13 @@
 #include "1_HardwareInputs.h"
 #include "2_HardwareOutputs.h"
+#include "3_ColorModeUtils.h"
 #include "4_ColorModes.h"
 #include "5_Debug.h"
 
 // Global variable inits
-uint16_t effectMode = 1;    // Chase, Fade
-uint16_t colorScheme = 1;   // Rainbow, Pairs
-uint16_t brightness = 1;    // 0 to 255
+uint8_t effectMode = 1;     // Chase, Fade
+uint8_t colorScheme = 1;    // Rainbow, Pairs
+uint8_t brightness = 1;     // 0 to 255
 uint16_t effectSpeed = 10;  // 0 to 1024
 bool tapEnabled = true;
